requirements.txt
google-genai
python-dotenv
streamlit
your_agent_code